+++ mosheh/doc.py
@@ -1,3 +1,15 @@
+"""
+Used to create the output documentation, this file deals with the codebase generated
+`custom_types.CodebaseDict` and creates `.md` files based on its collected information.
+
+The only public/exposed function here is `generate_doc`, which takes care of all of the
+private functions.
+
+There is a function for each step and statement type: `ast.Import`, `ast.ImportFrom`,
+`ast.Assign`, `ast.AnnAssign`, `ast.ClassDef`, `ast.FunctionDef`, `ast.AsyncFunctionDef`
+and `ast.Assert`, plus utility stuff like processing files.
+"""
+
 import subprocess
 from logging import Logger, getLogger
 from os import makedirs, path
@@ -236,12 +248,14 @@ def _codebase_to_markdown(filedata: list[StandardReturn], basedir: str) -> str:
     :rtype: str
     """
 
+    __meta__: StandardReturn = filedata.pop(0)
+
     filename: str = basedir.split(path.sep)[-1]
-    role: str = cast(FileRole, filedata.pop(0).get('__role__')).value
+    role: str = cast(FileRole, __meta__.get('__role__')).value
     filepath: str = (
         basedir.removesuffix(filename).replace(path.sep, '.').removesuffix('.')
     )
-    filedoc: str = ''
+    filedoc: str = cast(str, __meta__.get('__docstring__'))
     imports: str = ''
     constants: str = ''
     classes: str = ''
@@ -780,15 +794,15 @@ def _process_file(
     )
     folder_path: str = path.dirname(output_file_path)
 
-    if not path.exists(folder_path):
-        makedirs(folder_path)
+    if not path.exists(path.join('.', folder_path)):
+        makedirs(path.join('.', folder_path))
         logger.debug(f'{folder_path} created')
 
-    __write_to_file(output_file_path, content)
-    __update_navigation(folder_path, docs_path, key, output_file_path)
+    _write_to_file(output_file_path, content)
+    _update_navigation(folder_path, docs_path, key, output_file_path)
 
 
-def __write_to_file(file_path: str, content: str) -> None:
+def _write_to_file(file_path: str, content: str) -> None:
     """
     Writes content to a specified file.
 
@@ -814,12 +828,12 @@ def __write_to_file(file_path: str, content: str) -> None:
     :rtype: None
     """
 
-    with open(file_path, 'w', encoding='utf-8') as file:
+    with open(path.join('.', file_path), 'w', encoding='utf-8') as file:
         file.write(content)
         logger.debug(f'Content written to {file_path}')
 
 
-def __update_navigation(
+def _update_navigation(
     folder_path: str, docs_path: str, key: str, output_file_path: str
 ) -> None:
     """