+++ mosheh/handler.py
@@ -1,3 +1,13 @@
+"""
+Being the longest file, this one's role is process the source codebase.
+
+By calling `handle_def_nodes` with an `ast.AST` node, it's going to parse the node type
+and call the right handle func. The defined nodes are `ast.Import`, `ast.ImportFrom`,
+`ast.Assign`, `ast.AnnAssign`, `ast.FunctionDef`, `ast.AsyncFunctionDef`, `ast.ClassDef`
+and `ast.Assert`; if more nodes inside them, `_handle_node` is called to process the new
+one.
+"""
+
 import ast
 from logging import Logger, getLogger
 from typing import Final, cast