+++ mosheh/metadata.py
@@ -1,3 +1,9 @@
+"""
+If there is a file containing eater eggs and secrets, this one is not the one.
+
+Here are just info about Mosheh itself, data about data. No more than this.
+"""
+
 __author__ = 'LucasGonc'
 __copyright__ = 'Copyright (c) 2024 Lucas Gonçalves da Silva'
 __credits__ = ['LucasGonc']
@@ -6,6 +12,6 @@ __maintainer__ = 'LucasGonc'
 __email__ = 'lucasgoncsilva04@gmail.com'
 __repository__ = 'https://github.com/LucasGoncSilva/mosheh'
 __keywords__ = ['CLI', 'Python', 'documentation', 'MkDocs', 'automation', 'generation']
-__version__ = '1.3.3'
-__date__ = '2024-12-27'
+__version__ = '1.3.4'
+__date__ = '2025-01-07'
 __status__ = 'Production'